006/requirements.txt
beautifulsoup4==4.13.4
charset_normalizer==3.3.2
datasets==2.19.1
diskcache==5.6.3
fastapi==0.115.12
func_timeout==4.3.5
gradio==5.29.0
ipykernel==6.29.5
jieba==0.42.1
json5==0.9.14
jsonlines==4.0.0
jsonschema==4.23.0
jupyter_client==8.6.3
langchain==0.3.25
langchain_community==0.3.23
matplotlib==3.10.1
modelscope_studio==1.2.4
multiprocess==0.70.16
numpy==2.2.5
openai==1.77.0
optimum==1.24.0
pandas==2.2.3
pdfminer_six==20250416
pdfplumber==0.10.3
Pebble==5.1.0
Pillow==11.2.1
prettytable==3.16.0
pydantic==2.11.4
pytest==8.3.5
python_dateutil==2.8.2
python_docx==0.8.11
python_pptx==0.6.23
rank_bm25==0.2.2
regex==2023.10.3
Requests==2.32.3
scikit_learn==1.6.1
seaborn==0.13.2
setuptools==75.8.0
snowballstemmer==2.2.0
sympy==1.14.0
tabulate==0.9.0
tiktoken==0.9.0
timeout_decorator==0.5.0
torch==2.7.0
tqdm==4.67.1
transformers==4.49.0
uvicorn==0.34.2
